move test data to separate js file
