+++ client/src/assets/dummyStats.js
@@ -1,0 +1,79 @@
+"#d0a1ff"
+
+const keywordTestData = [
+    {
+        name: "Amazing day",
+        color: 'rgba(0,0,255,0.1)',
+        children: [
+            {
+                name: "Sunny day",
+                color: 'rgba(0,0,255,0.1)',
+                size: 340
+            },
+            {
+                name: "Passed test",
+                color: 'rgba(0,0,255,0.1)',
+                size: 640
+            },
+            {
+                name: "New car",
+                color: 'rgba(0,0,255,0.1)',
+                size: 1000
+            },
+            {
+                name: "Ate peperroni",
+                color: 'rgba(0,0,255,0.1)',
+                size: 920
+            },
+            {
+                name: "Drank a nice iced capp",
+                color: 'rgba(0,0,255,0.1)',
+                size: 920
+            },
+        ]
+    },
+    {
+        name: "okay day",
+        color: 'rgba(10,0,255,0.1)',
+        children: [
+            {
+                name: "Cloudy day",
+                color: 'rgba(10,0,255,0.1)',
+                size: 100
+            },
+            {
+                name: "did laundry",
+                color: 'rgba(10,0,255,0.1)',
+                size: 100
+            },
+            {
+                name: "nothing too special",
+                color: 'rgba(10,0,255,0.1)',
+                size: 100
+            },
+        ]
+    },
+    {
+        name: "bad day",
+        color: 'rgba(255,0,0,0.1)',
+        children: [
+            {
+                name: "rainy day",
+                color: 'rgba(255,0,0,0.1)',
+                size: 200
+            },
+            {
+                name: "failed my semester",
+                color: 'rgba(255,0,0,0.1)',
+                size: 700
+            },
+            {
+                name: "had breakup",
+                color: 'rgba(255,0,0,0.1)',
+                size: 1600
+            },
+        ]
+    },
+]
+
+export { keywordTestData }
